+++ change-log.md
@@ -1,4 +1,10 @@
 ### Change Log
 
+
+### 4.9.16 - 1.0.0
+- Client index view
+- Client create view
+- Client store method
+
 ### 4.9.16 - 1.0.0
 - Create clients table